[Performance] Extract PathSkipVoter to dedicated service and call before Filesystem::read()
